tests/wpt/web-platform-tests/old-tests/submission/Microsoft/sandbox/sandbox_023.htm

<!DOCTYPE HTML>
<html>
<head>
    <title>HTML5 Sandbox: Allow sandbox iframe to access other content from the same origin if sandbox="allow-same-origin".</title>
    <meta content="text/html; charset=UTF-8" http-equiv="Content-Type" />
    <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
    <link rel="help" href="http://dev.w3.org/html5/spec/Overview.html#sandboxed-origin-browsing-context-flag" />
    <meta name="assert" content=" Allow sandbox iframe to access other content from the same origin if sandbox='allow-same-origin'." />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <script type="text/javascript">
        
        
        var t = async_test("Allow sandbox iframe to access other content from the same origin if sandbox='allow-same-origin'");
  
        function callback(event)
        {
            t.step(function(){ 
                assert_true('sandbox' in document.createElement('iframe')); 
                assert_equals(event.data, "window.parent.document");
            });
            t.done();
        }

        var timer = setTimeout(callback, 4000);
        window.addEventListener("message", callback, false);
    </script>
    <iframe src="iframe_sandbox_023.htm" sandbox="allow-scripts allow-same-origin" style="display:none"></iframe>
    <div id=log></div>

</body>
</html>
